Recommendation_Sysyem/Recomdation_System/Recommendation/requirements.txt
mysql-connector-python==8.0.29
jsonschema==4.7.2
